Refactor: moved data_lifecycle and timestamps_extension
