+++ services/catalog/rs_server_catalog/app.py
@@ -16,17 +16,15 @@
 import asyncio
 import sys
 from contextlib import asynccontextmanager
-from http import HTTPStatus
 from os import environ as env
 from typing import Annotated
 
 import httpx
 from brotli_asgi import BrotliMiddleware
-from fastapi import Depends, FastAPI, HTTPException, Request, Security
-from fastapi.responses import JSONResponse
+from fastapi import Depends, FastAPI, Request, Security
 from fastapi.routing import APIRoute
 from httpx._config import DEFAULT_TIMEOUT_CONFIG
-from rs_server_catalog.data_lifecycle import DataLifecycle
+from rs_server_catalog.data_management.data_lifecycle import DataLifecycle
 from rs_server_catalog.middleware.catalog_middleware import CatalogMiddleware
 from rs_server_common import settings as common_settings
 from rs_server_common.authentication.apikey import (
@@ -42,15 +40,11 @@ from rs_server_common.middlewares import (
 from rs_server_common.settings import env_bool
 from rs_server_common.utils import init_opentelemetry
 from rs_server_common.utils.logging import Logging
-from stac_fastapi.api.errors import ErrorResponse
 from stac_fastapi.api.middleware import CORSMiddleware
 from stac_fastapi.pgstac.app import api
 from stac_fastapi.pgstac.app import app as sfpg_app
 from stac_fastapi.pgstac.app import with_transactions
 from stac_fastapi.pgstac.db import close_db_connection, connect_to_db
-from starlette.exceptions import HTTPException as StarletteHTTPException
-from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
-from starlette.responses import Response
 
 from .user_handler import CATALOG_PREFIX
