+++ services/catalog/tests/test_data_lifecycle.py
@@ -21,8 +21,8 @@ from datetime import datetime
 from urllib.parse import urlparse
 
 import pytest
-from rs_server_catalog.data_lifecycle import DataLifecycle
-from rs_server_catalog.timestamps_extension import ISO_8601_FORMAT
+from rs_server_catalog.data_management.data_lifecycle import DataLifecycle
+from rs_server_catalog.data_management.timestamps_extension import ISO_8601_FORMAT
 
 from tests.helpers import (
     TEMP_BUCKET,
